Add German extractor

+++ python/make_test_data.py
@@ -8,13 +8,13 @@ def main(args):
     for html_path, parsed in run_extractors_on_test_data(args.lang):
         json_path = html_path.replace('html', 'json')
         os.makedirs(os.path.dirname(json_path), exist_ok=True)
-        assert not os.path.exists(json_path)
-        with open(json_path, 'w') as json_file:
-            json.dump(
-                list(map(lambda item: item.convert(), parsed)),
-                json_file,
-                indent=4,
-                sort_keys=True)
+        if not os.path.exists(json_path):
+            with open(json_path, 'w') as json_file:
+                json.dump(
+                    list(map(lambda item: item.convert(), parsed)),
+                    json_file,
+                    indent=4,
+                    sort_keys=True)
 
 
 if __name__ == '__main__':
